change: created tiles; traps -> process_traps
